+++ Source/Tarkov-QuestTree-Server/ModInfo.cs
@@ -16,7 +16,7 @@ namespace QuestTreeServer
     /// </summary>
     public static class ModInfo
     {
-        public const string Version = "1.18.4";
+        public const string Version = "1.18.5";
 
         private static string? _stamp;