+++ Source/Tarkov-QuestTree/ModInfo.cs
@@ -13,7 +13,7 @@ namespace QuestTree
     /// </summary>
     internal static class ModInfo
     {
-        public const string Version = "1.18.4";
+        public const string Version = "1.18.5";
 
         private static string _stamp;